Added -f function to sh

+++ gb24-runner.sh
@@ -1,5 +1,23 @@
 #!/bin/bash
 
+BASE_PATH=$(pwd)
+
+echo "$BASE_PATH bp"
+
+if [[ " $@ " =~ " -h "  ]]; then
+  echo "gb24-runner help:"
+  echo "-----------------------"
+  echo "--clean - clean tmp folder"
+  echo "--pull - pull latest changes from git"
+  echo "  -f - hard reset to HEAD before pull"
+  echo "  -I <branch> - specify apI branch (default master)"
+  echo "  -P <branch> - specify apP branch (default master)"
+  echo "--gen-cert - generate 30-days self-signed SSL certificate"
+  echo "--no-daemon - do not run docker compose with -d flag (daemon)"
+  echo "--build - build all projects"
+  exit 0
+fi
+
 if [[ " $@ " =~ " --clean "  ]]; then
     echo ">>> Cleaning tmp..."
     rm -rf ./tmp
@@ -19,6 +37,19 @@ if [[ " $@ " =~ " --pull "  ]]; then
     esac 
     
   done
+  
+  if [[ " $@ " =~ " -f "  ]]; then
+    echo ">>> Resetting to HEAD..."
+    cd tmp/grzesbank_app && 
+    git reset --hard HEAD &&
+    git clean -f -d
+    cd $BASE_PATH
+    cd tmp/grzesbank-api &&
+    git reset --hard HEAD  &&
+    git clean -f -d
+    cd $BASE_PATH 
+  fi
+  
     echo ">>> Pulling from $BRANCH ..."
     cd tmp
     git clone -b $BRANCH_APP https://github.com/jewusiak/grzesbank_app || ( cd grzesbank_app && git pull && git checkout $BRANCH_APP && echo ">>> Pulled newest version of gb24-app" || echo ">>> Problem with gb24-app pull :(")
